backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic>=2.10.0
pydantic-settings>=2.7.0
httpx>=0.28.0
clerk-backend-api>=1.0.0
redis>=5.2.0
python-dotenv>=1.0.1
ruff>=0.8.0
google-generativeai>=0.8.0
openai>=1.60.0
